+++ json_serializable/test/test_files/json_test_example.non_nullable.g.dart
@@ -97,7 +97,8 @@ Numbers _$NumbersFromJson(Map<String, dynamic> json) => new Numbers()
       (json['doubles'] as List).map((e) => (e as num).toDouble()).toList()
   ..nnDoubles =
       (json['nnDoubles'] as List).map((e) => (e as num).toDouble()).toList()
-  ..duration = _fromJson(json['duration'] as int);
+  ..duration = _fromJson(json['duration'] as int)
+  ..date = _dateTimeFromEpochUs(json['date'] as int);
 
 abstract class _$NumbersSerializerMixin {
   List<int> get ints;
@@ -105,11 +106,13 @@ abstract class _$NumbersSerializerMixin {
   List<double> get doubles;
   List<double> get nnDoubles;
   Duration get duration;
+  DateTime get date;
   Map<String, dynamic> toJson() => <String, dynamic>{
         'ints': ints,
         'nums': nums,
         'doubles': doubles,
         'nnDoubles': nnDoubles,
-        'duration': _toJson(duration)
+        'duration': _toJson(duration),
+        'date': _dateTimeToEpochUs(date)
       };
 }